refactor(app): move layout to app and auth shell to shared ui
