+++ src/app/index.ts
@@ -1,2 +1,3 @@
 export { AppProviders } from './providers'
 export { AppRoutes } from './routes'
+export { Layout } from './Layout'
