--- src/components/index.ts
@@ -1,2 +1,0 @@
-export { AuthPage, AuthPageHeader } from './AuthPage'
-export { Layout } from './Layout'
