+++ src/modules/auth/LoginPage.tsx
@@ -1,7 +1,7 @@
 import { useId, useState } from 'react'
 import type { FormEvent } from 'react'
 import { Link, Navigate, useLocation, useNavigate } from 'react-router-dom'
-import { AuthPage, AuthPageHeader } from '@components/AuthPage'
+import { AuthPage, AuthPageHeader } from '@shared/ui'
 import { emailPattern, parseApiError } from '@shared/lib'
 import { FormNotice, TextInputField } from '@shared/ui'
 import { loginWithCredentials } from './client'
