+++ vite.config.ts
@@ -12,7 +12,6 @@ export default defineConfig({
     alias: {
       '@app': fileURLToPath(new URL('./src/app', import.meta.url)),
       '@auth': fileURLToPath(new URL('./src/modules/auth', import.meta.url)),
-      '@components': fileURLToPath(new URL('./src/components', import.meta.url)),
       '@lib': fileURLToPath(new URL('./src/lib', import.meta.url)),
       '@modules': fileURLToPath(new URL('./src/modules', import.meta.url)),
       '@shared': fileURLToPath(new URL('./src/shared', import.meta.url)),
